+++ src/assetbundles/src/js/TranslatorDetail.js
@@ -24,12 +24,9 @@ Craft.Translations.TranslatorDetail = {
         var $service = $('#service');
         var service = $service.val();
         var serviceValid = !!service;
-        var $sites = $(':checkbox[name="sites[]"]');
-        var sitesValid = $sites.filter(':checked').length > 0;
-        var valid = serviceValid && sitesValid;
+        var valid = serviceValid;
 
         this.toggleInputState($service, serviceValid, Craft.t('app', 'Please choose a translation service.'));
-        this.toggleInputState($sites, sitesValid, Craft.t('app', 'Please choose one or more sites.'));
 
         switch (service) {
             case 'acclaro':
@@ -58,19 +55,6 @@ Craft.Translations.TranslatorDetail = {
                 .appendTo($input)
                 .append($('<li>', {'text': message}));
         }
-    },
-
-    checkSelectedSites: function(sites) {
-        sites = JSON.parse(sites);
-        var $checkboxes = $(':checkbox[name="sites[]"]');
-        $checkboxes.each(function(){
-            var $el = $(this);
-            var value = $el.val();
-            if (sites.includes(value)) {
-                $el[0].defaultChecked = true;
-            }
-        });
-
     },
 
     serializeSettings: function(service) {
@@ -121,8 +105,12 @@ Craft.Translations.TranslatorDetail = {
             },
             function(data) {
                 if (data.success) {
-                    $('#status').val('active');
-                    Craft.cp.displayNotice(Craft.t('app', 'You are now authenticated!'));
+                    if (!data.error) {
+                        $('#status').val('active');
+                        Craft.cp.displayNotice(Craft.t('app', 'You are now authenticated!'));
+                    } else {
+                        Craft.cp.displayError(Craft.t('app', data.error));
+                    }
                 } else {
                     $('#status').val('inactive');
                     Craft.cp.displayError(Craft.t('app', 'Invalid API token.'));
@@ -138,11 +126,6 @@ Craft.Translations.TranslatorDetail = {
         $('#service').on('change', $.proxy(this.updateService, this));
 
         this.updateService();
-        
-        var sites = selectedSites;
-        if (sites !== '') {
-            this.checkSelectedSites(sites);
-        }
         
 
         $('.translations-authenticate-translation-service').on('click', function(e) {