+++ src/services/repository/TranslatorRepository.php
@@ -30,7 +30,6 @@ class TranslatorRepository
             'id',
             'label',
             'service',
-            'sites',
             'status',
             'settings',
         ]));
@@ -51,7 +50,6 @@ class TranslatorRepository
                 'id',
                 'label',
                 'service',
-                'sites',
                 'status',
                 'settings',
             ]));
@@ -76,7 +74,6 @@ class TranslatorRepository
                 'id',
                 'label',
                 'service',
-                'sites',
                 'status',
                 'settings',
             ]));
@@ -208,5 +205,21 @@ class TranslatorRepository
 
             throw $e;
         }
+    }
+
+    /**
+     * @return array id => label
+     */
+    public function getAcclaroApiTranslators()
+    {
+        $translators = array();
+
+        foreach ($this->getActiveTranslators() as $translator) {
+            if ($translator->service == 'acclaro') {
+                $translators[] = $translator->id;
+            }
+        }
+
+        return $translators;
     }
 }