+++ src/widgets/RecentOrders.php
@@ -10,6 +10,7 @@
 
 namespace acclaro\translations\widgets;
 
+use acclaro\translations\Translations;
 use Craft;
 use craft\db\Query;
 use craft\helpers\Db;
@@ -23,7 +24,8 @@ use acclaro\translations\records\WidgetRecord;
  * @package   Translations
  * @since     1.0.2
  */
-class RecentOrders extends Widget
+class
+RecentOrders extends Widget
 {
     /**
      * @inheritdoc
@@ -82,8 +84,10 @@ class RecentOrders extends Widget
         $view = Craft::$app->getView();
         
         $orders = $this->_getOrders();
+
+        $orderCountAcclaro = Translations::$plugin->orderRepository->getAcclaroOrdersCount();
         
-        return $view->renderTemplate('translations/_components/widgets/RecentOrders/body', ['orders' => $orders]);
+        return $view->renderTemplate('translations/_components/widgets/RecentOrders/body', ['orders' => $orders, 'orderCountAcclaro' => $orderCountAcclaro]);
     }
 
     public static function doesUserHaveWidget(string $type): bool